move bookCollection to a module
