+++ index.js
@@ -1,97 +1,7 @@
+import BookCollection from './modules/bookCollection.js';
+
 const bookListWrapper = document.getElementById('bookList');
 const addBookForm = document.getElementById('addBookForm');
-
-class BookCollection {
-  constructor(bookContainer) {
-    this.books = [];
-    this.bookContainer = bookContainer;
-    this.isStorageAvailable = false;
-    this.checkStorageAvailable('localStorage');
-
-    // get data from localStorage if exist!
-    this.initialDataFetchFromLocalStorage();
-
-    // add them to the page
-    this.books.forEach((book) => this.addBookToPage(book));
-  }
-
-  initialDataFetchFromLocalStorage() {
-    if (this.isStorageAvailable) {
-      const localData = window.localStorage.getItem('books');
-      if (localData) {
-        this.books = JSON.parse(localData);
-      }
-    }
-  }
-
-  checkStorageAvailable(type) {
-    let storage;
-    try {
-      storage = window[type];
-      const x = '__storage_test__';
-      storage.setItem(x, x);
-      storage.removeItem(x);
-      this.isStorageAvailable = true;
-    } catch (e) {
-      this.isStorageAvailable = false;
-    }
-  }
-
-  addBookToPage(data) {
-    const { id, title, author } = data;
-
-    this.bookContainer.innerHTML += `
-    <li id="${id}">
-      <h3>${title}</h3>
-      <p><em>By: ${author}</em></p>
-      <button class="removeBookBtn">Remove</button>
-    </li>
-    `;
-
-    // update event listener
-    this.updateEventListeners(this.bookContainer);
-  }
-
-  addBookToCollection(data) {
-    const { id, title, author } = data;
-
-    this.books.push({
-      id,
-      title,
-      author,
-    });
-
-    this.updateStorage();
-
-    // add to the page
-    this.addBookToPage(data);
-  }
-
-  updateEventListeners(element = document) {
-    const removeBookBtns = element.querySelectorAll('.removeBookBtn');
-
-    removeBookBtns.forEach((removeBtn) => {
-      removeBtn.addEventListener('click', (event) => {
-        const { parentNode } = event.target;
-        this.removeBook(parentNode.id);
-        parentNode.remove();
-      });
-    });
-  }
-
-  removeBook(id) {
-    this.books = this.books.filter((book) => book.id !== id);
-    this.updateStorage();
-  }
-
-  updateStorage() {
-    if (this.isStorageAvailable) {
-      const storage = window.localStorage;
-      storage.setItem('books', JSON.stringify(this.books));
-    }
-  }
-}
-
 const bookCollection = new BookCollection(bookListWrapper);
 
 addBookForm.addEventListener('submit', (event) => {
